MAINT: Remove abstract class as it bring no benefits
Moving comparison methods into their own modules. Removing abstract class as it has no benefit. Adjust tests for this change
