+++ .gitignore
@@ -2,4 +2,5 @@
 main_version
 branch_version
 src/__pycache__/*
+src/features/__pycache__/*
 tests/__pycache__/*
