+++ .pylintrc
@@ -5,9 +5,7 @@ max-line-length=120
 
 # Disable various warnings:
 
-# W0237 Disabled as it makes the code more readable
-# R0801 Disabled as it's a small amount of duplication
-disable=W0237, R0801
+disable=
 
 [MASTER]
 init-hook='import sys; sys.path.append("src")'
